% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{adelete}
\alias{adelete}
\title{~ *.zip - Suppression des fichiers en fin de traitement}
\usage{
adelete(finess, annee, mois, path, liste, type)
}
\arguments{
\item{finess}{Finess du fichier a supprimer}

\item{annee}{Annee du fichier}

\item{mois}{Mois du fichier}

\item{path}{Chemin d'acces aux fichiers}

\item{liste}{Liste des fichiers a effacer}

\item{type}{Type de fichier In / Out}
}
\description{
Supprime les fichiers de l'archive PMSI dezippes en début de traitement
}
\examples{
\dontrun{
   adelete(750712184,2016,2, path = '~/Exemple',  liste = c("rss","ano"), type = "in")
}

}
\seealso{
\code{\link{adezip}}, \code{\link{adezip2}}, \code{\link{astat}}
}
\author{
G. Pressiat
}
