% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_r.R
\name{ana_r_ca_gestes_marqueurs}
\alias{ana_r_ca_gestes_marqueurs}
\title{~ ANA - Chirurgie ambulatoire : 55 gestes marqueurs}
\usage{
ana_r_ca_gestes_marqueurs(
  p,
  periode = list(an = 2018, moi = 12),
  gestes_marqueurs
)
}
\arguments{
\item{p}{Noyau de paramètres}

\item{periode}{paramètres année et mois de l'envoi}

\item{gestes_marqueurs}{liste de requêtes des gestes marqueurs (nomensland)}
}
\value{
Taux de chirurgie ambulatoire et DMS > 0 nuit des gestes marqueurs par finess géographique
}
\description{
~ ANA - Chirurgie ambulatoire : 55 gestes marqueurs
}
\examples{
\dontrun{

library(dplyr, warn.conflicts = FALSE)
library(pmeasyr)

p <- noyau_pmeasyr(finess = '290000017',
                   annee  = 2018,
                   mois   = 12,
                   path   = '~/Documents/data/mco', 
                   tolower_names = TRUE,
                   n_max = Inf)


library(nomensland)

dicts <- get_dictionnaire_listes()
lgm <- get_all_listes('Chir ambu : 55 GM')

periodes <- list(
  list(an = 2013, moi = 12),
  list(an = 2014, moi = 12),
  list(an = 2015, moi = 12),
  list(an = 2016, moi = 12),
  list(an = 2017, moi = 12),
  list(an = 2018, moi = 12),
  list(an = 2019, moi = 11))

result <- periodes \%>\% purrr::map_dfr(ana_r_ca_gestes_marqueurs, p = p, gestes_marqueurs = lgm)
result <- result \%>\% arrange(`Geste marqueur`, `Période`)

pivot_result <- result \%>\% 
  select(`Geste marqueur`, nofiness, taux_ambu, `Nb total`, `Période`) \%>\% 
  mutate(stat = paste0(scales::percent(taux_ambu), ' (', `Nb total`, ')')) \%>\% 
  select(-taux_ambu, - `Nb total`) \%>\% 
  tidyr::spread(`Période`, stat, '')

}

}
