% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{astat}
\alias{astat}
\title{~ *.zip - Liste et volume des fichiers d'une archive PMSI}
\usage{
astat(path, file, view = T)
}
\arguments{
\item{path}{Chemin d'acces  a l'archive}

\item{file}{Nom du fichier archive}

\item{view}{par defaut a T : affiche la liste avec View(), a F retourne la table affichee a T}
}
\description{
Pour lister sans dezipper les fichiers d'une archive
}
\examples{
\dontrun{
   liste <- astat(path = '~/Documents/R/sources/2016/',
                  file = "750712184.2016.2.05042016093044.in.zip",
                  view = F)
}

}
\seealso{
\code{\link{adezip}}, \code{\link{adezip2}}
}
\author{
G. Pressiat
}
