% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R, R/rsf_before_utf.R
\name{collect_rsa_from_db}
\alias{collect_rsa_from_db}
\title{~ req : collecter les rsa présents dans une db}
\usage{
collect_rsa_from_db(con, an, n = Inf)

collect_rsa_from_db(con, an, n = Inf)
}
\arguments{
\item{con}{Connexion à la base de données}

\item{an}{année des rsa (ex: 17)}

\item{n}{Nombre de lignes à importer (équivalent `head`)}

\item{con}{Connexion à la base de données}

\item{an}{année des rsa (ex: 17)}

\item{n}{Nombre de lignes à importer (équivalent `head`)}
}
\value{
un liste rsa similaire à un import de pmeasyr

un liste rsa similaire à un import de pmeasyr
}
\description{
On selectionne certaines variables et on en cree d'autres utiles lors de l'execution de la requete

On selectionne certaines variables et on en cree d'autres utiles lors de l'execution de la requete
}
\examples{
\dontrun{
rsa <- collect_rsa_from_db(con, 16, n = 1e5)
}

\dontrun{
rsa <- collect_rsa_from_db(con, 16, n = 1e5)
}

}
\author{
G. Pressiat

G. Pressiat
}
