% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{creer_archive_vide}
\alias{creer_archive_vide}
\title{Créer une archive vide à partir d'une archive réelle}
\usage{
creer_archive_vide(chemin_archive, dossier_cible = "archive_simul/",
  finess_simul = "123456789")
}
\arguments{
\item{chemin_archive}{Chemin vers une archive réelle qui servira de modèle pour l'archive simulée}

\item{dossier_cible}{Chemin vers le dossier où l'archive simulée sera stockée}

\item{finess_simul}{Un numéro finess pour le fichier simulé}
}
\value{
\code{NULL}
}
\description{
Créer une archive vide à partir d'une archive réelle
}
