% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{creer_archives_vides}
\alias{creer_archives_vides}
\title{Creer des archives simulees a partir de plusieurs archives reelles}
\usage{
creer_archives_vides(chemins_archives)
}
\arguments{
\item{chemins_archives}{Chemin vers les archives qui servira de modèle pour l'archive simulée}
}
\description{
Creer des archives simulees a partir de plusieurs archives reelles
}
\seealso{
\code{\link[=creer_archive_vide]{creer_archive_vide()}}
}
\keyword{internal}
