% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{creer_json}
\alias{creer_json}
\title{~ req : creer un json pour partager une liste}
\usage{
creer_json(requete, chemin)
}
\arguments{
\item{requete}{`list` qui contient la requête}

\item{chemin}{Nom du fichier json qui sera créé}
}
\value{
un objet de classe rsa de pmeasyr
}
\description{
On selectionne certaines variables et on en cree d'autres utiles lors de l'execution de la requete
}
\examples{
\dontrun{
liste = list(nom_abrege = "pac",
             thematique = "test",
             actes = c('EBLA003', 'EQLF002'),
             dureemax = 0,
             auteur = "John Doe",
             date_saisie = "1960-01-01")
creer_json(liste, paste0('~/Documents/listes/', liste$nom_abrege, '.json'))
}

}
\author{
G. Pressiat
}
