% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{creer_nom_archive}
\alias{creer_nom_archive}
\title{Créer le nom d'une archive
Permet de simuler ou de créer des noms d'archives PMSI}
\usage{
creer_nom_archive(finess, annee, mois, type = "in", horodatage = NULL)
}
\arguments{
\item{finess}{Finess de l'archive.}

\item{annee}{Année de l'archive.}

\item{mois}{Mois de l'archive.}

\item{type}{Type de l'archive : \emph{in} ou \emph{out}.}

\item{horodatage}{Une date au format \emph{dmYHMS}}
}
\value{
Un nom d'archive compatible
}
\description{
Créer le nom d'une archive
Permet de simuler ou de créer des noms d'archives PMSI
}
