% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_generique}
\alias{db_generique}
\title{~ db - Copier un tibble dans une db}
\usage{
db_generique(con, an, table, prefix, suffix, indexes = list(), remove = T)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{an}{l'annee pmsi}

\item{table}{La table R (tibble) a copier dans la db}

\item{prefix}{prefixe de la table dans la db (ex : mco, rsf, ssr, ...)}

\item{suffix}{suffixe de la table dans la db (ex : rum_rum, rha_actes, rapss_rapss, ...)}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}
}
\value{
nothing
}
\description{
Copier une table R dans une db
}
\details{
La tables déjà importée dans R est copiée dans la db
}
\examples{
\dontrun{
purrr::quietly(db_generique)(con, 16, ma_table, 'had', 'rapss_ano') -> statuts ; gc(); #
# Result in db : had_16_rapss_ano
}
}
