% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_liste_tables}
\alias{db_liste_tables}
\title{~ db - Lister les tables d'une db en tableau}
\usage{
db_liste_tables(con, nb = 15)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{nb}{le nombre de lignes du tableau}
}
\value{
nothing
}
\description{
~ db - Lister les tables d'une db en tableau
}
\examples{
\dontrun{
db_liste_tables(con)
}
}
