% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_mco_in}
\alias{db_mco_in}
\title{~ db - Copier les rum dans une db}
\usage{
db_mco_in(con, 16, p)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{an}{l'annee pmsi}

\item{p}{le noyau pmeasyr}

\item{path1}{le chemin d'acces aux donnees}

\item{remove}{a TRUE, les tables precedentes rsa sont effacees avant}
}
\value{
nothing
}
\description{
Copier les rum, les actes et les diagnostics des rums dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db
La table diag est créée et la durée des rum est calculée (DUREESEJPART)
}
\examples{
\dontrun{
purrr::quietly(db_mco_in)(con, 16, p) -> statuts ; gc(); #ok
purrr::quietly(db_mco_in)(con, 15, p) -> statuts ; gc(); #..
purrr::quietly(db_mco_in)(con, 14, p) -> statuts ; gc(); #..
purrr::quietly(db_mco_in)(con, 13, p) -> statuts ; gc(); #..
purrr::quietly(db_mco_in)(con, 12, p) -> statuts ; gc(); #..
purrr::quietly(db_mco_in)(con, 11, p) -> statuts ; gc(); #..
}
}
