% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_ssr_out}
\alias{db_ssr_out}
\title{~ db - Copier les rha dans une db}
\usage{
db_ssr_out(con, p, remove = T, zip = T, indexes = list(), ...)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rha de l'annee sont effacees avant}

\item{zip}{a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}
}
\value{
nothing
}
\description{
Copier les rha, les actes, les diagnostics des rha, ssrha et ano dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db.
Le tra est ajouté aux tables.
}
\examples{
\dontrun{
purrr::quietly(db_ssr_out)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_ssr_out)(con, p, annee = 2017, mois = 7) -> statuts ; gc(); #..
}
}
