% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{dico}
\alias{dico}
\title{~ Dico - Dictionnaire des tables}
\usage{
dico(table)
}
\arguments{
\item{table}{Table dont on veut le dictionnaire de variables}
}
\description{
Obtenir le dictionnaire d'une table
}
\examples{
\dontrun{
# N'importer qu'une ligne du fichier :
   irsa('750712184', 2016, 8, '~/path/path', typi= 1, n_max = 1) -> import
   dico(import$rsa)
}

}
\seealso{
\code{\link{irsa}}, \code{\link{irum}}
}
\author{
G. Pressiat
}
