% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{enrobeur}
\alias{enrobeur}
\title{~ req : mise en forme d'une liste de codes}
\usage{
enrobeur(a, robe = "'", colonne = F, interstice = ", ",
  symetrique = F)
}
\description{
~ req : mise en forme d'une liste de codes
}
\examples{
\dontrun{
li <- c('QEFA003', 'QEFA005', 'QEFA010', 'QEFA013', 'QEFA015', 'QEFA019', 'QEFA020')

enrobeur(li, robe="", interstice="|") \%>\% cat()

enrobeur(li, robe="\\'", interstice=",") \%>\% cat()

enrobeur(li, robe="\\'\%", interstice="\\n", symetrique = T) \%>\% cat()
}

}
\author{
G. Pressiat
voir \url{https://guillaumepressiat.shinyapps.io/transcodeur/} pour son utilisation interactive hors AP-HP
}
