% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{epmsi_mco_rae}
\alias{epmsi_mco_rae}
\title{~ VVR - Reproduire le tableau RAE}
\usage{
epmsi_mco_rae(valo, knit = FALSE)
}
\arguments{
\item{valo}{Un tibble résultant de \code{\link{vvr_mco}}}

\item{knit}{à TRUE, appliquer une sortie `knitr::kable` au résultat}
}
\value{
Un tibble similaire au tableau RAE epmsi
}
\description{
~ VVR - Reproduire le tableau RAE
}
\examples{
\dontrun{
epmsi_mco_rae(valo)
}

}
\author{
G. Pressiat
}
