% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{epmsi_mco_rav}
\alias{epmsi_mco_rav}
\title{~ VVR - Reproduire le tableau RAV}
\usage{
epmsi_mco_rav(valo)
}
\value{
Un tibble similaire au tableau RAV epmsi
}
\description{
~ VVR - Reproduire le tableau RAV
}
\examples{
\dontrun{
epmsi_mco_rav(valo)
}

}
\author{
G. Pressiat
}
