% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{epmsi_mco_rav2}
\alias{epmsi_mco_rav2}
\title{~ VVR - Reproduire le tableau RAV (works)}
\usage{
epmsi_mco_rav2(valo, theorique = TRUE)
}
\arguments{
\item{valo}{Un tibble résultant de \code{\link{vvr_mco}}}

\item{knit}{à TRUE, appliquer une sortie `knitr::kable` au résultat}
}
\value{
Un tibble similaire au tableau RAV epmsi (montant BR* ou BR si coeff prud = 1)
}
\description{
~ VVR - Reproduire le tableau RAV (works)
}
\examples{
\dontrun{
epmsi_mco_rav2(valo)
}

}
\author{
G. Pressiat
}
