% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{epmsi_mco_rav_rum}
\alias{epmsi_mco_rav_rum}
\title{~ VVR - Synthese des rubriques de valo apres valo des RUM}
\usage{
epmsi_mco_rav_rum(valo_rum, theorique = TRUE)
}
\arguments{
\item{valo_rum}{Un tibble résultant de \code{\link{vvr_rum}}}
}
\value{
Un tibble similaire au tableau RAV epmsi
}
\description{
~ VVR - Synthese des rubriques de valo apres valo des RUM
}
\examples{
\dontrun{
epmsi_mco_rav_rum(valo)
}

}
\seealso{
epmsi_mco_rav, epmsi_mco_rav_rum, epmsi_mco_rav2,
}
\author{
G. Pressiat
}
