% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmsi_mco_vmed_f.R
\name{epmsi_mco_vmed_f}
\alias{epmsi_mco_vmed_f}
\title{~ MCO - Comprendre le tableau epmsi 1.V.2.VMED - F}
\usage{
epmsi_mco_vmed_f(p, ref_indic)
}
\arguments{
\item{p}{Noyau de paramètres}

\item{ref_indic}{table des indications ATIH}
}
\value{
Liste de deux tables : tableau ePMSI et détail avec NAS, UCD, Indcations, etc.
}
\description{
~ MCO - Comprendre le tableau epmsi 1.V.2.VMED - F
}
\examples{
\dontrun{
library(pmeasyr)
library(dplyr, warn.conflicts = FALSE)


p <- noyau_pmeasyr(
  finess = '123456789',
  annee  = 2019,
  mois   = 3,
  path   = '~/Documents/data/mco',
  tolower_names = TRUE,
  lib = FALSE, 
  progress = FALSE
)

# devtools::install_github('GuillaumePressiat/nomensland')
library(nomensland)
ref_indic <- get_table('mco_medref_atih_indications') \%>\% 
  mutate(mois = substr(anseqta,1,2), 
         annee = substr(anseqta, 3,6)) 
vmed_f <- epmsi_mco_vmed_f(p, 
                           ref_indic)
}

}
