% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{extraire_types_fichiers}
\alias{extraire_types_fichiers}
\title{Extraire les types de fichiers
Permet d'afficher les types de fichiers d'une archive (ano, rss, rsa, ...)}
\usage{
extraire_types_fichiers(selection_fichiers_a_extraire, info_archive)
}
\arguments{
\item{selection_fichiers_a_extraire}{Fichiers sélectionnés dans le zip (résultat de selectionne_fichiers)}

\item{info_archive}{informatin de l'archive (résultat de selectionne_archive)}
}
\value{
un vecteur
}
\description{
Extraire les types de fichiers
Permet d'afficher les types de fichiers d'une archive (ano, rss, rsa, ...)
}
