% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iano_mco}
\alias{iano_mco}
\title{~ MCO - Import des Anohosp}
\usage{
iano_mco(finess, annee, mois, path, lib = T, tolower_names = F, typano = "out")
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des données sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{typano}{Type de donnees In / Out}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame ou tbl_df) qui contient les données Anohosp in / out
}
\description{
Import du fichier ANO In ou Out.
}
\details{
Formats depuis 2011 pris en charge
Structure du nom du fichier attendu  :
\emph{finess.annee.moisc.ano}
\emph{finess.annee.moisc.ano.txt}

\strong{750712184.2016.2.ano}
\strong{750712184.2016.2.ano.txt}
}
\examples{
\dontrun{
   iano_mco('750712184',2015,12,'~/Documents/data/mco') -> ano_out15
   iano_mco('750712184',2015,12,'~/Documents/data/mco', typano = "in") -> ano_in15
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
