% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iano_psy}
\alias{iano_psy}
\title{~ PSY - Import des Anohosp}
\usage{
iano_psy(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tbl_df) contenant les données Anohosp SSR du Out.
}
\description{
Import du fichier Ano Out
}
\details{
Formats depuis 2012 pris en charge
Structure du nom du fichier attendu (sortie de Genrha) :
\emph{finess.annee.moisc.ano}

\strong{750712184.2016.2.ano}
}
\examples{
\dontrun{
   anoh <- iano_psy('750712184',2015,12,"~/Documents/data/psy")
}

}
\seealso{
\code{\link{irpsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
