% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iano_rafael}
\alias{iano_rafael}
\title{~ RSF - Import des Anohosp RSFA}
\usage{
iano_rafael(finess, annee, mois, path, lib = T, lamda = F, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lamda}{a TRUE, importe le fichier ano-ace-maj}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame ou tbl_df) qui contient les données Anohosp in / out
}
\description{
Import du fichier ANO-ACE RSF Out ou le ano-ace-maj (reprise)
}
\details{
Formats depuis 2012 pris en charge pour les ano-ace
Formats depuis 2014 pris en charge pour les ano-ace-maj (reprise 2013)

Structure du nom du fichier attendu  :
\emph{finess.annee.moisc.ano}

\strong{750712184.2016.2.ano}
}
\examples{
\dontrun{
   iano_rafael('750712184', 2015, 12,'~/Documents/data/rsf') -> ano_out15
   iano_rafael('750712184', 2015, 12,'~/Documents/data/rsf', lamda = T) -> lamda_maj_ano_out14
}

}
\seealso{
\code{\link{irafael}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
