% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iium_corresp}
\alias{iium_corresp}
\title{~ MCO - Import des donnees correspondances d'UM du Out}
\usage{
iium_corresp(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2020)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires à passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tibble) contenant les informations Um out/UM in.
}
\description{
Imports du fichier IUM corresp MCO
}
\details{
Formats depuis 2020 pris en charge
}
\examples{
\dontrun{
   um_corresp <- iium_corresp('750712184',2020,12,"~/Documents/data/mco")
}

}
\seealso{
\code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
