% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iivg_mco}
\alias{iivg_mco}
\title{~ MCO - Import des IVG}
\usage{
iivg_mco(finess, annee, mois, path, lib = T, tolower_names = F, typivg = c('out', 'in'))
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{typmed}{Type de donnees In / Out}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tibble) contenant les IVG In ou Out 
(si le fichier n'existe pas, pas de donnée importée). 
Le TYPEPREST est par defaut 53
}
\description{
Import des fichiers IVG In ou Out.
}
\details{
Formats depuis 2017 pris en charge
}
\examples{
\dontrun{
   iivg_mco('750712184', 2015, 12,'~/Documents/data/mco') -> ivg_out15
   iivg_mco('750712184', 2015, 12,'~/Documents/data/mco', typivg = "in") -> ivg_in15
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat, N. Taright
}
