% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{ileg_mco}
\alias{ileg_mco}
\title{~ MCO - Import des erreurs Leg}
\usage{
ileg_mco(finess, annee, mois, path, reshape = F, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{reshape}{booleen TRUE/FALSE : la donnee doit-elle etre restructuree ? une ligne = une erreur, sinon, une ligne = un sejour. par defaut a F}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tbl_df) contenant les erreurs Out.
}
\description{
Import de la liste d'erreurs de génération Genrsa
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   ileg_mco('750712184',2015,12,'~/Documents/data/mco') -> leg15
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
