% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{ileg_ssr}
\alias{ileg_ssr}
\title{~ SSR - Import des erreurs Leg}
\usage{
ileg_ssr(finess, annee, mois, path, reshape = F, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{reshape}{booleen TRUE/FALSE : la donnee doit-elle etre restructuree ? une ligne = une erreur, sinon, une ligne = un sejour. par defaut a F}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}
}
\value{
Une table (data.frame, tbl_df) contenant les erreurs Out.
}
\description{
Import de la liste d'erreurs de génération Genrha
}
\examples{
\dontrun{
   ileg_had('750712184',2015,12,'~/Documents/data/ssr') -> leg15
}

}
\seealso{
\code{\link{irha}}, \code{\link{issrha}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
