% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{imed_mco}
\alias{imed_mco}
\title{~ MCO - Import des Med}
\usage{
imed_mco(finess, annee, mois, path, typmed = c("out", "in"), lib = T, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{typmed}{Type de donnees In / Out}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a TRUE ; necessite le package \code{sjmisc}}

\item{...}{parametres supplementaires a passer
dans la fonction \code{readr::read_fwf()}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tbl_df) contenant les médicaments In ou Out (T2A, ATU et thrombo selon l'existence des fichiers : si le fichier n'existe pas, pas de donnée importée). Pour discriminer le type de prestation, la colonne TYPEPREST donne l'information : T2A 06 - ATU 09 - THROMBO 10
}
\description{
Import des fichiers MED In ou Out.
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   imed_mco(750712184,2015,12,'~/pathpath/') -> med_out15
   imed_mco(750712184,2015,12,'~/pathpath/', typmed = "in") -> med_in15
}

}
\author{
G. Pressiat
}
\seealso{
irum irsa
}

