% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{inner_tra}
\alias{inner_tra}
\title{~ TRA - Ajout du TRA aux donnees Out}
\usage{
inner_tra(table, tra, sel = 1, champ = "mco")
}
\arguments{
\item{table}{Table a laquelle rajouter le tra}

\item{tra}{tra a rajouter}

\item{sel}{Variable a garder du tra ; sel = 1 : numero de sejour, sel = 2 : toutes les variables}

\item{champ}{Champ PMSI : mco, had, ssr, psy : deux tra en psy : psy_rpsa, psy_r3a}
}
\value{
Une table contenant le inner_join entre table et tra
}
\description{
Ajout du TRA par dplyr::inner_join
}
\examples{
\dontrun{
   med <- imed_mco('750712184',2015,12,"~/Documents/data/mco","out")
   tra <- itra('750712184',2015,12,"~/Documents/data/mco")
   med <- inner_tra(med,tra)
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}}, \code{\link{imed_mco}}, \code{\link{irpsa}}, \code{\link{irha}}, \code{\link{irapss}}
}
\author{
G. Pressiat
}
