% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{ipo}
\alias{ipo}
\title{~ MCO - Import des PO}
\usage{
ipo(finess, annee, mois, path, typpo = c("out", "in"), lib = T, ...)
}
\arguments{
\item{...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}

\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{typpo}{Type de donnees In / Out}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjmisc}}
}
\value{
Une table (data.frame, tibble) contenant les prélèvements d'organes In ou Out.
}
\description{
Imports des fichiers PO In / Out
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   po <- ipo(750712184,2015,12,"~/Documents/data/mco")
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}}, utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
