% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{irha}
\alias{irha}
\title{~ SSR - Import des RHA}
\usage{
irha(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du fichier Out de GENRHA a integrer}

\item{annee}{Annee de la periode (du fichier Out)}

\item{mois}{Mois de la periode (du fichier Out)}

\item{path}{Chemin d'acces au fichier .rha}

\item{lib}{Attribution de libelles aux colonnes}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max=10e3} pour lire les 1000 premieres lignes}
}
\description{
Import des RHA
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   irha('750712184',2015,12,'pathpath/') -> rha15
}
}
\seealso{
\code{\link{iano_ssr}}, \code{\link{ileg_ssr}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
