% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs.R
\name{irhs}
\alias{irhs}
\title{~ SSR - Import des RHS}
\usage{
irhs(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du fichier In de GENRHA a integrer}

\item{annee}{Annee de la periode (du fichier in)}

\item{mois}{Mois de la periode (du fichier in)}

\item{path}{Chemin d'acces au fichier .rhs.rtt}

\item{lib}{Attribution de libelles aux colonnes}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max=10e3} pour lire les 10000 premieres lignes}
}
\description{
Import des RHS
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   irhs('750712184',2015,12,'pathpath/') -> rhs15
}
}
\seealso{
\code{\link{iano_ssr}}, \code{\link{ileg_ssr}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
