% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{irsa}
\alias{irsa}
\title{~ MCO - Import des RSA}
\usage{
irsa(finess, annee, mois, path, lib = T, typi = 4, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{typi}{Type d'import, par defaut a 4, a 0 : propose a l'utilisateur de choisir au lancement}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une classe S3 contenant les tables (data.frame, tbl_df ou tbl) importées  (rsa, rsa_um, actes et das si import > 3)
}
\description{
Import des RSA. 6 types d'imports possibles.
}
\details{
Formats depuis 2011 pris en charge
Structure du nom du fichier attendu (sortie de Genrsa) :
\emph{finess.annee.moisc.rsa}

\strong{750712184.2016.2.rsa}

Types d'imports :
\tabular{ll}{
1 Light : \tab partie fixe (très rapide)\cr
2 Light+ : \tab Partie fixe + stream en ligne (+) actes et das\cr
3 Light++ : \tab Partie fixe + stream en ligne (++) actes, das, typaut um et dpdr des um\cr
4 Standard : \tab Partie fixe + création des tables acdi et rsa_um\cr
5 Standard+ : \tab Partie fixe + création des tables acdi et rsa_um + stream (+)\cr
6 Standard++ : \tab Partie fixe + création des tables acdi et rsa_um + stream (++)\cr
}

\strong{Principe du streaming :}
Mise en chaîne de caractères de la succession d'actes CCAM au cours du RUM, par exemple, pour un RUM :
\samp{"ACQK001, LFQK002, MCQK001, NAQK015, PAQK002, PAQK900, YYYY600, ZZQP004"}

La recherche d'un (ou d'une liste d') acte(s) sur un RUM est largement accélérée, comparée à une requête sur la large table acdi par une requête du type :

\code{grepl("ZZQP004",rsa$actes)}  # toutes les lignes de RSA avec au moins un ZZQP004

\code{e66 <- grepl('E66',das)|grepl('E66',dpdrum)}  # toutes les lignes de RSA avec un diagnostic E66

Cela permet de n'utiliser que la seule table rsa avec stream et d'avoir les infos sur les séjours directement : nb séjours, journées, entrée / sortie (...) plutôt que d'avoir à utiliser et croiser les tables acdi, rsa_um avec rsa.
}
\examples{
\dontrun{
   irsa('750712184',2015,12,'~/Documents/data/mco') -> rsa15
}

}
\seealso{
\code{\link{irum}}, \code{\link{ileg_mco}}, \code{\link{iano_mco}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
