% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{irum}
\alias{irum}
\title{~ MCO - Import des RUM}
\usage{
irum(finess, annee, mois, path, lib = T, typi = 3, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du In a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{typi}{Type d'import, par defaut a 3, a 0 : propose a l'utilisateur de choisir au lancement}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une classe S3 contenant les tables (data.frame, tibble) importées (rum, actes, das et dad si import 3 et 4)
}
\description{
Import des RUM. 4 types d'imports possibles.
}
\details{
Formats depuis 2011 pris en charge

Structure du nom du fichier attendu (entrée pour Genrsa) :
\emph{finess.annee.moisc.rum}

\strong{750712184.2016.2.rum}

Types d'imports :
\tabular{ll}{
1 XLight : \tab partie fixe\cr
2 Light : \tab partie fixe +  streaming des actes, dad et das\cr
3 Standard : \tab partie fixe + table acdi\cr
4 Standard+ : \tab Import standard (3) + stream\cr
}

\strong{Principe du streaming :}
Mise en chaîne de caractères de la succession d'actes CCAM au cours du RUM, par exemple, pour un RUM :
\samp{"ACQK001, LFQK002, MCQK001, NAQK015, PAQK002, PAQK900, YYYY600, ZZQP004"}

La recherche d'un (ou d'une liste d') acte(s) sur un RUM est largement accélérée, comparée à une requête sur la large table acdi par une requête du type :

\code{grepl("ZZQP004",rum$actes)}  # toutes les lignes de  RUM avec au moins un ZZQP004
\code{grepl("ZZQP004|EBLA003",rum$actes)}  # toutes les lignes de  RUM avec au moins un ZZQP004 ou un EBLA003
}
\examples{
\dontrun{
   irum('750712184',2015,12,'~/Documents/data/mco', typi = 1) -> rum15
}

}
\seealso{
\code{\link{irsa}}, \code{\link{ileg_mco}}, \code{\link{iano_mco}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
