% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{issrha}
\alias{issrha}
\title{~ SSR - Import des SSRHA}
\usage{
issrha(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles a la table : T}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tbl_df) contenant les données SHA, et a partir de 2017 une liste de deux tables (sha et gme)
}
\description{
Import du fichier SHA
}
\details{
Formats depuis 2011 pris en charge
Structure du nom du fichier attendu (sortie de Genrha) :
\emph{finess.annee.moisc.sha}

\strong{750712184.2016.2.sha}
}
\examples{
\dontrun{
   sha <- issrha('750712184',2015,12,"~/Documents/data/ssr")
}

}
\seealso{
\code{\link{irha}}, \code{\link{ileg_ssr}}, \code{\link{iano_ssr}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
