% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{itra}
\alias{itra}
\title{~ TRA - Import du TRA}
\usage{
itra(finess, annee, mois, path, lib = T, tolower_names = F, champ = "mco")
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{champ}{Champ PMSI du TRA a integrer ("mco", "ssr", "had", "psy_rpsa", ", "psy_r3a"), par defaut "mco"}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premières lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame ou tbl_df) qui contient : - Clé RSA - NORSS - Numéro de ligne du fichier RSS d'origine (rss.ini) - NAS - Date d'entrée du séjour - GHM groupage du RSS (origine) - Date de sortie du séjour
}
\description{
Import du fichier TRA, 4 champs PMSI couverts.
}
\details{
Formats depuis 2011 pris en charge
Structure du nom du fichier attendu (sortie de Genrsa) :
\emph{finess.annee.moisc.tra}

\strong{750712184.2016.2.tra}
}
\examples{
\dontrun{
   itra('750712184',2015,12,'~/Documents/data/champ_pmsi') -> tra15
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}}, \code{\link{ileg_mco}}, \code{\link{iano_mco}}, \code{\link{irha}}, \code{\link{irapss}}, \code{\link{irpsa}}, \code{\link{ir3a}}, 
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
