% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{labeleasier}
\alias{labeleasier}
\title{~ Labels pour le PMSI}
\usage{
labeleasier(col, Mode_entree = F, Mode_sortie = F, Provenance = F,
  Destination = F, Sexe = F, facteur = F)
}
\arguments{
\item{col}{Colonne à laquelle attribuer le libellé}

\item{Mode_entree}{'6' : 'Mutation'}

\item{Mode_sortie}{'9' : 'Décès'}

\item{Provenance}{'1' : 'MCO'}

\item{Destination}{'6' : 'HAD'}

\item{Sexe}{'2' : 'Femme'}
}
\value{
Un vecteur caractère ou facteur
}
\description{
Attribuer des libelles aux colonnes PMSI
}
\examples{
\dontrun{
   labeleasier(rsa$rsa$SEXE, Sexe = T, F)
   labeleasier(rsa$rsa$DEST, Destination = T, F)
}

}
\author{
G. Pressiat
}
