% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{lancer_requete}
\alias{lancer_requete}
\title{~ req : lancer une ou plusieurs requetes avec une ou des listes}
\usage{
lancer_requete(tables, elements, vars = NULL)
}
\value{
un tibble concatenant les resultats de toutes les requetes : les rsa qui ont un acte de la liste, un diag, une duree correspondante, etc.
}
\description{
~ req : lancer une ou plusieurs requetes avec une ou des listes
}
\examples{
\dontrun{
prepare_rsa(rsa) -> rsa
get_all_listes('Recours Exceptionnel') -> listes_re
lancer_requete(rsa, listes_re)

get_liste('chir_bariatrique_total') -> liste_bari
lancer_requete(rsa, liste_bari)
}

}
\author{
G. Pressiat
}
