% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{noyau_pmeasyr}
\alias{noyau_pmeasyr}
\alias{print.pm_param}
\title{~ par - Noyau de parametres}
\usage{
noyau_pmeasyr(...)

\method{print}{pm_param}(p)
}
\description{
Définir un noyau de paramètres
}
\details{
Voir exemple
}
\examples{
\dontrun{
library(magrittr)

p <- noyau_pmeasyr(
finess = '750712184',
annee  = 2016,
mois   = 12,
path   = '~/Documents/data/mco'
)

c(p, type = "out", liste = "") \%>\% adezip()

p \%>\% irsa()     -> rsa
p \%>\% iano_mco() -> ano
p \%>\% ipo()      -> po

c(p, type = "in", liste = "") \%>\% adezip()
p \%>\% irum()     -> rum

# Modifier le type d'import :
c(p, typi = 6) \%>\% irsa() -> rsa

# Pour visualiser p : 
p
print(p)
}

}
\author{
G. Pressiat
}
