% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{prepare_rsa}
\alias{prepare_rsa}
\title{~ req : preparer les rsa pour la requete}
\usage{
prepare_rsa(rsa)
}
\value{
un objet de classe rsa de pmeasyr
}
\description{
On selectionne certaines variables et on en cree d'autres utiles lors de l'execution de la requete
}
\examples{
\dontrun{
tab <- irsa(typi = 6)
prepare_rsa(tab)
}

}
\author{
G. Pressiat
}
