% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renomme_fichier.R
\name{renomme_fichier}
\alias{renomme_fichier}
\title{Renommer un fichier au format standardisé}
\usage{
renomme_fichier(finess = "000000000", annee, mois, path, nom_fichier,
  type_fichier, garder_original = TRUE, copie_temporaire = TRUE,
  dossier_cible = path)
}
\arguments{
\item{nom_fichier}{Nom du fichier original à renommer}

\item{type_fichier}{Extension du fichier. Par exemple 'rss.txt'.}

\item{garder_original}{Si `TRUE` alors le fichier original n'est pas modifié. Sinon, il est renommé.}

\item{copie_temporaire}{Si `TRUE` alors le fichier renommé est copié dans un répertoire temporaire. Si `FALSE`, alors le nouveau fichier est  Cet argument est pris en compte que si garder_original == TRUE est valable.}

\item{dossier_cible}{Dossier où la copie renommée du fichier doit être enregistré. Valable que si garder_original == TRUE.}
}
\value{
Chemin vers le fichier renommé.
}
\description{
Cette fonction permet de renommer un fichier, 
par exemple directement produit à partir du système d'information (hospitalier), sous une forme standardisée
afin de pouvoir l'importer avec pmeasyr tout en gardant optionnellement le fichier original.
}
\examples{
\dontrun{
library(pmeasyr)
# classique
renomme_fichier(annee = 2019, 
                mois = 12, 
                path = '~/Documents/data/mco/',
                nom_fichier = 'monrss.txt', 
                type_fichier = 'rss.txt', 
                garder_original = TRUE)

# avec noyau de paramètres
p <- noyau_pmeasyr(
  finess = '123456789',
  annee  = 2019,
  mois   = 12,
  path   = '~/Documents/data/mco',
  tolower_names = TRUE,
  lib = FALSE, 
  progress = FALSE
)

renomme_fichier(p, 
                nom_fichier = 'monrss.txt', 
                type_fichier = 'rss.txt', 
                garder_original = TRUE)

}
}
