% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{requete}
\alias{requete}
\title{~ req : requeter les rsa avec une liste}
\usage{
requete(tables, elements, vars = NULL)
}
\value{
un tibble contenant les rsa respectant la requete : les rsa qui ont un acte de la liste, un diag, une duree correspondante, etc.
}
\description{
~ req : requeter les rsa avec une liste
}
\examples{
\dontrun{
liste = list(actes = c('EBLA003', 'EQLF002'))
requete(rsa, liste)

liste = list(actes = c('EBLA003', 'EQLF002'), dureemax = 0)
requete(rsa, liste)
}

}
\author{
G. Pressiat
}
