% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{requete_db}
\alias{requete_db}
\title{~ req : requeter les rsa dans une db avec une liste}
\usage{
requete_db(con, an, elements, vars = NULL)
}
\value{
un tibble contenant les rsa respectant la requete : les rsa qui ont un acte de la liste, un diag, une duree correspondante, etc.
}
\description{
~ req : requeter les rsa dans une db avec une liste
}
\examples{
\dontrun{
liste = list(actes = c('EBLA003', 'EQLF002'))
requete_db(con, an, liste)

liste = list(actes = c('EBLA003', 'EQLF002'), dureemax = 0)
requete_db(con, 16, liste)
}

}
\author{
G. Pressiat
}
