% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{selectionne_fichiers}
\alias{selectionne_fichiers}
\title{Selectionne les fichiers dans une archive}
\usage{
selectionne_fichiers(chemin_archive, types_fichier)
}
\value{
Le noms dans fichiers dans l'archive
}
\description{
Selectionne les fichiers dans une archive
}
