% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{silence}
\alias{silence}
\title{~ Silence - import sans messages}
\usage{
silence(fonction)
}
\arguments{
\item{fonction}{Fonction a modifier}
}
\value{
La fonction qui n'imprime rien dans la console
}
\description{
Fonction d'import silencieuse avec la fonction \code{purrr::quietly}
}
\examples{
\dontrun{
   silence(irsa()) -> rsa_silence
   # Table RSA : 
   rsa_silence(750712184, 2017, 1, '~/data/mco', typi = 1)$result$rsa -> rsa17
}

}
\author{
G. Pressiat
}

