% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_had}
\alias{tbl_had}
\title{remote access aux tables had}
\usage{
tbl_had(con, 16, table)
}
\value{
tibble
}
\description{
remote access aux tables had
}
\examples{
\dontrun{
tbl_had(con, 16, 'rapss_rapss')
}
}
