% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_mco}
\alias{tbl_mco}
\title{~ db - remote access aux tables mco}
\usage{
tbl_mco(con, 16, table)
}
\value{
tibble
}
\description{
~ db - remote access aux tables mco
}
\examples{
\dontrun{
tbl_mco(con, 16, 'rsa_rsa')
}
}
