% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_psy}
\alias{tbl_psy}
\title{~ db - remote access aux tables psy}
\usage{
tbl_psy(con, an, table)
}
\value{
tibble
}
\description{
~ db - remote access aux tables psy
}
\examples{
\dontrun{
tbl_psy(con, 16, 'rpsa_rpsa')
}
}
