% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R, R/rsf_before_utf.R
\name{tbl_rsf}
\alias{tbl_rsf}
\title{~ db - remote access aux tables rsf}
\usage{
tbl_rsf(con, an, table)

tbl_rsf(con, an, table)
}
\arguments{
\item{con}{Connexion à la base de données}

\item{an}{Année pmsi (ex: 16)}

\item{table}{Table à requêter}
}
\value{
remote table

remote table
}
\description{
~ db - remote access aux tables rsf

~ db - remote access aux tables rsf
}
\examples{
\dontrun{
tbl_rsf(con, 16, 'rsf_rafael_ano')
}
\dontrun{
tbl_rsf(con, 16, 'rsf_rafael_ano')
}
}
