% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_ssr}
\alias{tbl_ssr}
\title{~ db - remote access aux tables ssr}
\usage{
tbl_ssr(con, 16, table)
}
\value{
tibble
}
\description{
~ db - remote access aux tables ssr
}
\examples{
\dontrun{
tbl_ssr(con, 16, 'rha_rha')
}
}
