% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{vvr_mco}
\alias{vvr_mco}
\title{~ VVR - Croiser parties GHS / supplements et ano}
\usage{
vvr_mco(rsa_v, ano_sv)
}
\arguments{
\item{rsa_v}{tibble résultant de la fonction \code{\link{vvr_rsa}}}

\item{ano_sv}{tibble résultant de la fonction \code{\link{vvr_ano_mco}}}
}
\value{
Un tibble final contenant la catégorie du tableau SV epmsi, et les variables rec_ de recette par séjour
}
\description{
On ajoute les données facturation (vvr_mco_sv) aux données de valorisation 100% T2A (vvr_ghs_supp)
}
\details{
C'est un left join
}
\examples{
\dontrun{
}

}
\author{
G. Pressiat
}
