% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{vvr_mco_sv}
\alias{vvr_mco_sv}
\title{~ VVR - Attribuer le caractere facturable par cle_rsa}
\usage{
vvr_mco_sv(rsa, ano, porg = tibble(cle_rsa = ""))
}
\value{
Un tibble contenant la catégorie du tableau SV epmsi, une ligne par clé rsa
}
\description{
Reproduire les catégories du tableau SV d'epmsi, à partir des tables résultant des fonctions
\code{\link{vvr_rsa}}, \code{\link{vvr_ano_mco}} et éventuellement d'une table contenant le fichcomp PO
}
\examples{
\dontrun{
# Tenir compte des porg
vvr_mco_sv(vrsa, vano, porg = ipo(p))

# ne pas tenir compte des porg
vvr_mco_sv(vrsa, vano)
}

}
\seealso{
\code{\link{vvr_ano_mco}}, \code{\link{vvr_rsa}}, \code{\link{vvr_mco}}
}
\author{
G. Pressiat
}
