% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{vvr_rsa_hors_periode}
\alias{vvr_rsa_hors_periode}
\title{~ VVR - Forcer le groupage des RSA hors période en erreur}
\usage{
vvr_rsa_hors_periode(vrsa, an_v, mois_v)
}
\description{
Pour les RSA transmis hors période, on modifie CMD, GHM, et GHS en 90, 90Z99Z, et 9999 pour ne pas les valoriser
}
\examples{
\dontrun{
noyau_pmeasyr(
  finess = '290000017',
  annee  = 2019,
  mois   = 11,
  path   = '~/Documents/data/mco',
  progress = FALSE,
  lib = FALSE, 
  tolower_names = TRUE
) -> p

vrsa <- vvr_rsa(p)
vrsa <- vrsa \%>\% 
  vvr_rsa_hors_periode(as.character(p$annee), stringr::str_pad(p$mois, 2, "left", '0'))
}

}
\seealso{
\code{\link{epmsi_mco_sv}}, \code{\link{vvr_rsa}}, \code{\link{vvr_mco}}
}
\author{
G. Pressiat
}
