% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{vvr_rum}
\alias{vvr_rum}
\title{~ VVR - Distribuer la valorisation des rsa au niveau des rum ou des passages UM}
\usage{
vvr_rum(
  p,
  valo,
  repartition_multi = "{prop_pmct_um}*0.5+{prop_pass}*0.5",
  pmct_mono = c(FALSE, TRUE),
  seuil_pmct = 10,
  type_passage = c("RUM", "Passage unique")
)
}
\arguments{
\item{p}{Un noyau de paramètres}

\item{valo}{Un tibble résultant de \code{\link{vvr_mco}}}

\item{repartition_multi}{pour renseigner les paramètres de la clef de répartition (entre durée de passage et PMCT des UM fréquentées)}

\item{pmct_mono}{Calcul du PMCT par UM sur les mono-RUM (TRUE), ou sur tous les séjours par l'UM fournissant le DP (FALSE)}

\item{seuil_pmct}{En dessous de quel nombre on considère le PMCT non robuste, dans ce cas, on passe à une distribution uniquement sur les durées de passages}

\item{type_passage}{La table résultat est soit au niveau RUM, soit au niveau passage unique (pas de UM A, UM B, UM A, juste UM A, UM B)}
}
\value{
Un tibble
}
\description{
~ VVR - Distribuer la valorisation des rsa au niveau des rum ou des passages UM
}
\examples{
\dontrun{
vvr_rum(p, valo, type_passage = "RUM", pmct_mono = FALSE)
}

}
\seealso{
epmsi_mco_rav_rum, vvr_mco
}
\author{
G. Pressiat

fbrcdnj
}
