% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{adelete}
\alias{adelete}
\title{~ *.zip - Suppression des fichiers en fin de traitement}
\usage{
adelete(finess, annee, mois, path, liste, type, ...)
}
\arguments{
\item{finess}{Finess du fichier a supprimer}

\item{annee}{Annee du fichier}

\item{mois}{Mois du fichier}

\item{path}{Chemin d'acces aux fichiers}

\item{liste}{Liste des fichiers a effacer : par defaut a "", efface tous les \code{fichiers finess.annee.mois.}}

\item{type}{Type de fichier In / Out : par defaut a "", efface tous les fichiers \code{finess.annee.mois.}}
}
\description{
Supprime les fichiers de l'archive PMSI dezippes en debut de traitement
}
\examples{
\dontrun{
   adelete('750712184',2016,2, path = '~/Exemple',  liste = c("rss","ano"), type = "in")
   
   adelete('750712184',2016,2, path = '~/Exemple')
}

}
\seealso{
\code{\link{adezip}}, \code{\link{adezip2}}, \code{\link{astat}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
