% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{adezip}
\alias{adezip}
\title{~ *.zip - Dezippe des fichiers de l'archive PMSI}
\usage{
adezip(finess, annee, mois, path, liste, pathto = "", type, recent = T)
}
\arguments{
\item{finess}{Finess du fichier a dezipper}

\item{annee}{Annee du fichier}

\item{mois}{Mois du fichier}

\item{path}{Chemin d'acces au fichier}

\item{liste}{des fichiers à dezipper ex: ano, rss, rsa, dmi, ... ; si liste = "", dezippe la totalite de l'archive}

\item{type}{Type de l'archive : in / out}

\item{recent}{par defaut a T, l'archive la plus recente sera utilisee, sinon propose a l'utilisateur de choisir quelle archive dezipper}

\item{pathto}{par defaut a "", dezipper la ou est l'archive, sinon preciser le chemin ou dezipper les fichiers (ailleurs)}
}
\description{
Dezipper une archive PMSI au besoin
}
\examples{
\dontrun{
     adezip(750712184,2016,2, path = '~/Documents/R/sources/2016',
            liste = 'med',
            pathto = "~/Exemple",
            type = "out")

     adezip(750712184,2016,2, path = '~/Documents/R/sources/2016',
            liste = c('med','rapss', 'ano'),
            pathto = "~/Exemple",
            type = "in")

     adezip(750712184,2016,2, path = '~/Documents/R/sources/2016',
            liste = c('rss', 'ano'),
            pathto = "~/Exemple",
            type = "in",
            recent = F)
}

}
\seealso{
\code{\link{adezip2}}, \code{\link{astat}}, \code{\link{adelete}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
