% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{adezip}
\alias{adezip}
\alias{adezip.pm_param}
\alias{adezip.default}
\title{~ *.zip - Identifie et dézippe des fichiers de l'archive PMSI}
\usage{
adezip(...)

\method{adezip}{pm_param}(.params, ...)

\method{adezip}{default}(finess, annee, mois, path, liste = NULL,
  pathto = path, type = "in", recent = TRUE, nom_archive = NULL)
}
\arguments{
\item{...}{Paramètres supplémentaires. Permet par exemple de changer un des paramètres après avoir passé un noyau de paramètres sans changer le noyau de paramètres.}

\item{.params}{Un noyau de paramètres définis par la fonction fonction \code{\link{noyau_pmeasyr}}}

\item{finess}{Finess de l'archive.}

\item{annee}{Année de l'archive.}

\item{mois}{Mois de l'archive.}

\item{path}{Chemin d'accès au répertoire contenant l'archive à décompresser}

\item{liste}{Vecteur de caractère avec le type de fichiers à dézipper (ex: ano, rss, rsa, dmi, ...). Par défaut, \code{liste = ""} dezippe la totalite de l'archive.}

\item{pathto}{Par defaut la même valeur que \code{path}, dézipper dans le même répertoire que l'archive, sinon préciser le chemin ou dezipper les fichiers dans le répertoire indiqué par \code{pathto}.}

\item{type}{Type de l'archive : \emph{in} ou \emph{out}.}

\item{recent}{par défaut \code{TRUE}, l'archive la plus recente sera utilisee, sinon propose à l'utilisateur de choisir quelle archive dezipper}

\item{nom_archive}{Nom de l'archive à décompresser dans le dossier \code{path}. Par défaut, \code{NULL}, n'utilise pas ce paramètre. Si le chemin est spécifié, alors les paramètres \code{finess}, \code{annee}, \code{mois} et \code{recent} ne sont pas utilisés.}
}
\value{
Les chemins d'accès des fichiers décompressés, de manière invisible.
}
\description{
Recherche et dézippe (décompresse) les fichiers contenus dans une archive \emph{*.in} ou \emph{*.out} du PMSI en fonction de paramètres. Il est possible de passer directement les paramètres permettant d'identifier l'archive à dézipper (méthode par défaut) ou à l'aide de paramètres enregistrés dans un noyau de paramètres (voir fonction \code{\link{noyau_pmeasyr}}).

Dézipper les fichiers sélectionnés d'une archive
}
\examples{
\dontrun{
     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = 'med',
            pathto = "~/Exemple",
            type = "out")

     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = c('med','rapss', 'ano'),
            pathto = "~/Exemple",
            type = "in")

     adezip('750712184',2016,2, path = '~/Documents/R/sources/2016',
            liste = c('rss', 'ano'),
            pathto = "~/Exemple",
            type = "in",
            recent = F)
            
     # Utilisation avec un noyau de paramtères
     p <- noyau_pmeasyr(
            finess = '750712184',
            annee  = 2016,
            mois   = 12,
            path   = '~/Documents/data/mco',
            progress = F
            )
     
     adezip(p, type = "in")
     
     # Modification d'un paramètre du noyay
     adezip(p, mois = 11, type = "in")
}

}
\seealso{
\code{\link{adezip2}}, \code{\link{astat}}, \code{\link{adelete}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}

\code{\link[utils:unzip]{utils::unzip()}}
}
\author{
G. Pressiat
}
