% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{astat}
\alias{astat}
\title{~ *.zip - Liste et volume des fichiers d'une archive PMSI}
\usage{
astat(path, file, view = TRUE)
}
\arguments{
\item{path}{Chemin d'accès au dossier contenant les archives}

\item{file}{Nom du fichier d'archive}

\item{view}{par défaut \code{TRUE} : affiche la liste avec \code{\link[utils:View]{utils::View()}}  Sinon retourne la table dans la console.}
}
\value{
Une \code{data.frame} contenant les colonnes :
\itemize{
\item \emph{Name} Nom du fichier
\item \emph{Date} Date de modification
\item \emph{Taille (Mo)}
}
}
\description{
Pour lister sans dezipper les fichiers d'une archive.
}
\examples{
test_files_dir <- system.file("extdata", "test_data", "test_adezip", package = "pmeasyr")
astat(path = test_files_dir, file = "123456789.2016.2.15032016152413.in.zip", view = FALSE)

}
\seealso{
\code{\link[=adezip]{adezip()}} et \code{\link[=adezip2]{adezip2()}} permettent de décompresser les fichiers. Cette fonction est basée sur \code{\link[utils:unzip]{utils::unzip()}} en utilisant l'argument \code{list = TRUE}
}
\author{
G. Pressiat
}
