% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{creer_archives_vides}
\alias{creer_archives_vides}
\title{Créer des archives simulées à partir de plusieurs archives réelles}
\usage{
creer_archives_vides(chemins_archives)
}
\arguments{
\item{chemins_archives}{Chemin vers les archives qui servira de modèle pour l'archive simulée}
}
\value{
\code{NULL}
}
\description{
Créer des archives simulées à partir de plusieurs archives réelles
}
\seealso{
\code{\link[=creer_archive_vide]{creer_archive_vide()}}
}
