% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_mco_in}
\alias{db_mco_in}
\title{~ db - Copier les rum dans une db}
\usage{
db_mco_in(con, p, remove = T, zip = T, indexes = list(), ...)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rum de l'annee sont effacees avant}

\item{zip}{a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}
}
\value{
nothing
}
\description{
Copier les rum, les actes et les diagnostics des rums dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db
La table diag est créée et la durée des rum est calculée (DUREESEJPART)
}
\examples{
\dontrun{
purrr::quietly(db_mco_in)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_mco_in)(con, p, annee = 2015) -> statuts ; gc(); #..
}
}
