% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_mco_out}
\alias{db_mco_out}
\title{~ db - Copier les rsa dans une db}
\usage{
db_mco_out(con, p, remove = T, zip = T, indexes = list(),  ...)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rsa de l'annee sont effacees avant}

\item{zip}{a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}
}
\value{
nothing
}
\description{
Copier les rsa, les passages um, les actes et les diagnostics des rsa, et ano dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db.
La table diag est créée, les variables ghm, année séquentielle des tarifs
et un champ caractère diagnostics sont ajoutés à la table rsa.
Le tra est ajouté aux tables.
}
\examples{
\dontrun{
purrr::quietly(db_mco_out)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_mco_out)(con, p, annee = 2017, mois = 7) -> statuts ; gc(); #..
}
}
