% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R, R/rsf_before_utf.R
\name{db_rsf_out}
\alias{db_rsf_out}
\title{~ db - Copier les rsf dans une db}
\usage{
db_rsf_out(con, p, remove = T, zip = T, indexes = list(), ...)

db_rsf_out(con, p, remove = T, zip = T, indexes = list(), ...)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rafael de l'annee sont effacees avant}

\item{zip}{a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}

\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rafael de l'annee sont effacees avant}

\item{zip}{a TRUE les fichiers des archives sont dezippes et effaces apres integration dans la db}

\item{indexes}{index a ajouter a la table dans la base (voir \code{\link[dplyr]{copy_to}})}
}
\value{
nothing

nothing
}
\description{
Copier tous les rsf (lettre par lettre, A, B, C, ...), les ano-ace dans une db

Copier tous les rsf (lettre par lettre, A, B, C, ...), les ano-ace dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db

Les tables sont importées dans R puis copiées dans la db
}
\examples{
\dontrun{
purrr::quietly(db_rsf_out)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_rsf_out)(con, p, annee = 2014) -> statuts ; gc(); #ok
}
\dontrun{
purrr::quietly(db_rsf_out)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_rsf_out)(con, p, annee = 2014) -> statuts ; gc(); #ok
}
}
