% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{db_rsf_out}
\alias{db_rsf_out}
\title{~ db - Copier les rsf dans une db}
\usage{
db_rsf_out(con, p, remove = T, ...)
}
\arguments{
\item{con}{la connexion a la base de donnees (src_..)}

\item{p}{le noyau pmeasyr}

\item{remove}{a TRUE, les tables precedentes rsa sont effacees avant}
}
\value{
nothing
}
\description{
Copier tous les rsf (lettre par lettre, A, B, C, ...), les ano-ace dans une db
}
\details{
Les tables sont importées dans R puis copiées dans la db
}
\examples{
\dontrun{
purrr::quietly(db_rsf_out)(con, p) -> statuts ; gc(); #ok
purrr::quietly(db_rsf_out)(con, p, annee = 2014) -> statuts ; gc(); #ok
}
}
