% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{epmsi_mco_sv}
\alias{epmsi_mco_sv}
\title{~ VVR - Reproduire le tableau SV}
\usage{
epmsi_mco_sv(valo)
}
\value{
Un tibble similaire au tableau SV epmsi
}
\description{
~ VVR - Reproduire le tableau SV
}
\examples{
\dontrun{
epmsi_mco_sv(valo)
}

}
\author{
G. Pressiat
}
