% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{iium}
\alias{iium}
\title{~ MCO - Import des donnees UM du Out}
\usage{
iium(finess, annee, mois, path, lib = T, tolower_names = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires à passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tbl_df) contenant les informations structures du Out.
}
\description{
Imports du fichier IUM MCO
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   um <- iium('750712184',2015,12,"~/Documents/data/mco")
}

}
\seealso{
\code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
