% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{imed_mco}
\alias{imed_mco}
\title{~ MCO - Import des Med}
\usage{
imed_mco(finess, annee, mois, path, lib = T, tolower_names = F, typmed = c('out', 'in'))
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{typmed}{Type de donnees In / Out}

\item{lib}{Ajout des libelles de colonnes aux tables, par defaut a \code{TRUE} ; necessite le package \code{sjlabelled}}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{parametres supplementaires a passer
dans la fonction \code{\link[readr]{read_fwf}}, par exemple
\code{n_max = 1e3} pour lire les 1000 premieres lignes,  \code{progress = F, skip = 1e3}}
}
\value{
Une table (data.frame, tibble) contenant les médicaments In ou Out 
(T2A, ATU et thrombo selon l'existence des fichiers : si le fichier n'existe pas, pas de donnée importée). 
Pour discriminer le type de prestation, la colonne TYPEPREST donne l'information : T2A 06 - ATU 09 - THROMBO 10
}
\description{
Import des fichiers MED In ou Out.
}
\details{
Formats depuis 2011 pris en charge
}
\examples{
\dontrun{
   imed_mco('750712184',2015,12,'~/Documents/data/mco') -> med_out15
   imed_mco('750712184',2015,12,'~/Documents/data/mco', typmed = "in") -> med_in15
}

}
\seealso{
\code{\link{irum}}, \code{\link{irsa}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
