% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{irafael}
\alias{irafael}
\title{~ RSF - Import des Rafael}
\usage{
irafael(finess, annee, mois, path, lib = T, tolower_names = F, 
stat = T, lister = c("A", "B",
"C", "H", "L", "M", "P"), lamda = F, ...)
}
\arguments{
\item{finess}{Finess du Out a importer : dans le nom du fichier}

\item{annee}{Annee PMSI (nb) des donnees sur 4 caracteres (2016)}

\item{mois}{Mois PMSI (nb) des donnees (janvier : 1, decembre : 12)}

\item{path}{Localisation du fichier de donnees}

\item{lib}{Ajout des libelles de colonnes aux tables, par défaut a TRUE ; necessite le package \code{sjlabelled}}

\item{stat}{avec stat = T, un tableau synthetise le nombre de lignes par type de rafael}

\item{lister}{Liste des types d'enregistrements a importer}

\item{lamda}{a TRUE, importe les fichiers \code{rsfa-maj} de reprise de l'annee passee}

\item{tolower_names}{a TRUE les noms de colonnes sont tous en minuscules}

\item{~...}{Autres parametres a specifier \code{n_max = 1e3}, ...}
}
\value{
Une classe S3 contenant les tables (data.frame, tbl_df ou tbl) importées  (rafaels)
}
\description{
Import des Rafael et des Rafael reprises
}
\details{
Formats depuis 2012 pour les rsfa
Formats depuis 2014 pour les rsfa-maj (reprise 2013)
}
\examples{
\dontrun{
   irafael('750712184',2015,12,'~/Documents/data/rsf') -> rsfa15
   irafael('750712184',2015,12,'~/Documents/data/rsf', lister = 'C', lamda = T) -> rsfa14_lamda
}

}
\seealso{
\code{\link{iano_rafael}},
utiliser un noyau de parametres avec \code{\link{noyau_pmeasyr}}
}
\author{
G. Pressiat
}
