% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{parse_nom_fichier}
\alias{parse_nom_fichier}
\title{Decoupe un seul nom de fichier}
\usage{
parse_nom_fichier(nom_fichier, format_date_archive = "\%d\%m\%Y\%H\%M\%S")
}
\arguments{
\item{nom_fichier}{Le nom du fichier à découper. Chaine de caractère de longueur 1.}

\item{format_date_archive}{Format de date d'horodatage pour les fichiers archive avec la notation de \code{\link[base:strptime]{base::strptime()}}.}
}
\description{
Decoupe un seul nom de fichier
}
\seealso{
\code{\link[=parse_noms_fichiers]{parse_noms_fichiers()}} pour traiter plusieurs noms de fichiers.
}
\keyword{internal}
