% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archives.R
\name{parse_nom_fichier}
\alias{parse_nom_fichier}
\title{Extraire les informations d'un nom de fichier}
\usage{
parse_nom_fichier(nom_fichier, format_date_archive = "\%d\%m\%Y\%H\%M\%S")
}
\arguments{
\item{nom_fichier}{Une chaine de caractères du fichier à découper}
}
\value{
Une liste avec les informations extraites du nom de fichier :
\itemize{
\item \code{nom_fichier} Le nom de fichier passé en arguement
\item \code{finess}
\item \code{annee}
\item \code{mois}
\item \code{horodatage_production} L'horodatage de production pour les fichiers \emph{in} et \emph{out}
\item \code{type} Type de fichier : \emph{in}, \emph{out}, \emph{rss}...
}
}
\description{
Extraire les informations d'un nom de fichier
}
