% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_had}
\alias{tbl_had}
\title{~ db - remote access aux tables had}
\usage{
tbl_had(con, an, table)
}
\value{
tibble
}
\description{
~ db - remote access aux tables had
}
\examples{
\dontrun{
tbl_had(con, 16, 'rapss_rapss')
}
}
