% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tbl_had}
\alias{tbl_had}
\title{~ db - remote access aux tables had}
\usage{
tbl_had(con, an, table)
}
\arguments{
\item{con}{Connexion à la base de données}

\item{an}{Année pmsi (ex: 16)}

\item{table}{Table à requêter}
}
\value{
remote table

tibble
}
\description{
~ db - remote access aux tables had
}
\examples{
\dontrun{
tbl_had(con, 16, 'rapss_rapss')
}
}
