% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_before_utf.R
\name{tdiag}
\alias{tdiag}
\title{~ Tidy - Tidy Diagnostics}
\usage{
tdiag(d, include = T)
}
\arguments{
\item{d}{Objet S3 resultat de l'import pmeasyr (irsa, irum, irha)}

\item{include}{booleen : defaut a T; T : restructure l'objet S3 (agglomere dp, dr, das et dad, par exemple)}
}
\description{
Restructurer les diagnostics
}
\details{
On obtient une table contenant tous les diagnostics par séjour, sur le principe suivant :
Une variable numérique indique la position des diagnostics
- pour les rsa : 1 : DP du rsa, 2 : DR du rsa, 3 : DPUM, 4 : DRUM, 5 : DAS
- pour les rum : 1 : DP du rum, 2 : DR du rum, 3 : DAS, 4 : DAD
- pour les rha : 1 : MMP du rha, 2 : FPPC du rha, 3 : AE, 4 : DA
}
\examples{
\dontrun{
# avec include = T
irum('750712184', 2016, 8, '~/path/path', typi = 3) -> d1
tdiag(d1) -> d1
d1$diags
d1$actes
d1$dad
irsa('750712184', 2016, 8, '~/path/path', typi = 4) -> d1
tdiag(d1, include = F) -> alldiag
}

}
\seealso{
\code{\link{irsa}}, \code{\link{irum}}, \code{\link{irha}}
}
\author{
G. Pressiat
}
