% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VVR_MCO.R
\name{vvr_ghs_supp}
\alias{vvr_ghs_supp}
\title{~ VVR - Attribuer les recettes GHS et suppléments sur des rsa}
\usage{
vvr_ghs_supp(rsa, tarifs, supplements = NULL, ano = tibble(cle_rsa = ""),
  porg = tibble(cle_rsa = ""), diap = tibble(cle_rsa = ""), full = FALSE,
  cgeo = 1.07, prudent = NULL, bee = TRUE)
}
\value{
Un tibble contenant les différentes rubriques de valorisation, une ligne par clé rsa
}
\description{
Reproduire la valorisation BR et coefficients géo/prudentiels du tableau RAV d'epmsi, à partir des tables résultant des fonctions
\code{\link{vvr_rsa}}, \code{\link{vvr_ano_mco}}, et de tables contenant les fichcomp PO et DIAP
}
\details{
Pour l'heure cette fonction ne tient pas compte de fichcomp PIE
}
\examples{
\dontrun{
# Récupérer les tarifs GHS et des suppléments (ex-DGF) : 
tarifs      <- nomensland::get_table('tarifs_mco_ghs') \%>\% distinct(ghs, anseqta, .keep_all = TRUE)
supplements <- nomensland::get_table('tarifs_mco_supplements')

# Recette GHS de base et suppléments EXB, EXH
vvr_ghs_supp(vrsa, tarifs)

# Recette GHS de base et suppléments EXB, EXH, et des suppléments (hors PIE)
vvr_ghs_supp(vrsa, tarifs, supplements, vano, ipo(p), idiap(p), bee = FALSE)
}

}
\seealso{
\code{\link{vvr_ano_mco}}, \code{\link{vvr_rsa}}, \code{\link{vvr_mco}}
}
\author{
G. Pressiat
}
